Revert "Revert "Switching to multiple environment variables.""
This reverts commit 635a969c53a5ff5a1bf49b89c8f4a9fbcfbe92c8.
